requirements.txt
absl-py==2.2.2
altair==5.5.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arviz==0.21.0
astropy==6.1.7
astropy-iers-data==0.2025.5.26.0.37.21
astroquery==0.4.10
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autograd==1.7.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
blackjax==1.2.5
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
cons==0.4.6
contourpy==1.3.2
corner==2.2.3
cycler==0.12.1
Cython==3.0.12
dask==2025.4.0
dataclasses==0.6
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deneb==1.0.1
diffeqtorch==1.0.0
distributed==2025.4.0
dm-tree==0.1.9
elfi==0.8.7
equinox==0.12.1
etils==1.12.2
etuples==0.3.9
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
fastprogress==1.0.3
filelock==3.18.0
flatbuffers==25.2.10
flatlatex==0.15
flax==0.10.6
flowjax==17.1.1
fonttools==4.57.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
GPy==1.13.2
grpcio==1.71.0
h11==0.16.0
h5netcdf==1.6.1
h5py==3.13.0
healpy==1.18.1
html5lib==1.1
humanize==4.12.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipyparallel==9.0.1
ipython==8.35.0
ipython-genutils==0.2.0
jabbar==0.0.16
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jax==0.5.0
jax-sgmc @ git+https://github.com/mattocallaghan/jax-sgmc.git@e8659cb7ceeae3113fbb75395a2a20962979823a
jaxlib==0.5.0
jaxopt==0.8.5
jaxtyping==0.3.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
julia==0.6.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.9.2
keyring==25.6.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
locket==1.0.0
logical-unification==0.4.6
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
miniKanren==1.0.3
mistune==3.1.3
ml_collections==1.1.0
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multipledispatch==1.0.0
namex==0.0.9
narwhals==1.36.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nflows==0.14
notebook==6.4.12
numba==0.61.2
numdifftools==0.9.41
numpy==1.26.4
numpyro==0.18.0
opt_einsum==3.4.0
optax==0.2.4
optree==0.15.0
orbax-checkpoint==0.11.12
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
paramax==0.0.0
paramz==0.9.6
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
portpicker==1.6.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyabc==0.12.15
pyabcranger==0.0.72
pycparser==2.22
pyerfa==2.0.1.5
Pygments==2.19.1
pyknos==0.16.0
pymc==5.22.0
pyparsing==3.2.3
pyro-api==0.1.2
pyro-ppl==1.9.1
PySocks==1.7.1
pystellibs @ git+https://github.com/mfouesneau/pystellibs.git@f32f89a999c2a9f29261a59f33665d4695f7d897
pytensor==2.30.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyvo==1.7
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
sbi==0.24.0
sbibm==1.1.0
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
selenium==4.31.0
Send2Trash==1.8.3
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.13.1
tblib==3.1.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
tensorstore==0.1.74
termcolor==3.0.1
terminado==0.18.1
tf_keras==2.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
toolz==1.0.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.9
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vega-datasets==0.9.0
wadler_lindig==0.1.5
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xarray==2025.3.1
xarray-einstats==0.8.0
zict==3.0.0
zipp==3.21.0
zuko==1.4.0
